% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasca.R
\name{canvasca}
\alias{canvasca}
\title{Generating abstract art from cellular automata}
\usage{
canvasca(colors, n = 100, iterations = 20, probability = 0.5, conway = FALSE)
}
\arguments{
\item{n}{number of rows and columns}

\item{iterations}{number of iterations of the algorithm}

\item{probability}{the probability of a starting sell being "alive"}

\item{conway}{a boolean for if Conway's Game of Life algorithm will be used}

\item{color}{a vector of two colors to be used for the art}
}
\value{
a ggplot of abstract art
}
\description{
Generating abstract art from cellular automata
}
