% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvaslogreg.R
\name{canvaslogreg}
\alias{canvaslogreg}
\title{Generating abstract art from logistic regression}
\usage{
canvaslogreg(color1, color2, n = 100, resolution = 30, rdist, ...)
}
\arguments{
\item{color1}{first color hexcode}

\item{color2}{second color hexcode}

\item{n}{number of generated observations for the test data}

\item{resolution}{resolution of generated image}

\item{rdist}{function for generating values (i.e. runif, rnorm, rexp)}

\item{...}{other arguments passed into the function that generates values}
}
\value{
a ggplot of abstract art
}
\description{
Generating abstract art from logistic regression
}
