% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvastree.R
\name{canvastree}
\alias{canvastree}
\title{Generating abstract art from a decision tree}
\usage{
canvastree(colors, n = 100, resolution = 30, rdist, ...)
}
\arguments{
\item{colors}{a vector of color hex codes}

\item{n}{number of generated observations for the test data}

\item{resolution}{resolution of generated image}

\item{rdist}{function for generating values (i.e. runif, rnorm, rexp)}

\item{...}{other arguments passed into the function that generates values}
}
\value{
a ggplot of abstract art
}
\description{
Generating abstract art from a decision tree
}
