% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logfile_existence.R
\name{check_logfile_existence}
\alias{check_logfile_existence}
\title{Check, if a logfile exists}
\usage{
check_logfile_existence()
}
\arguments{
\item{file_name}{the file name for saving the plot}
}
\value{
a data frame
}
\description{
This function loads the logfile, if it exists. If it doesn't, it will create one.
}
\examples{
check_logfile_existence()
}
\seealso{
\code{\link{generate_logfile_entry}} get the data to be logged
}
