% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasca.R
\name{game_of_life}
\alias{game_of_life}
\title{Run Conway's Game of Life algorithm on a matrix}
\usage{
game_of_life(mat, iterations)
}
\arguments{
\item{mat}{a matrix}

\item{iterations}{number of iterations of the algorithm}
}
\value{
resulting matrix from the algorithm
}
\description{
Run Conway's Game of Life algorithm on a matrix
}
\references{
\url{https://www.youtube.com/watch?v=6byzbkDe-RI}
}
