% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(formula)
}
\arguments{
\item{formula}{a list that contains formulas for transforming the x- and y-values.}
}
\value{
data frame
}
\description{
The generative images are based on values in a dataframe. This function creates the data by transforming the base values `seq(-pi, pi)` with a `formula`.
}
\examples{
\dontrun{
generate_data(formula)
}
# an example for a formula:
formula <- list(
  x = quote(runif(1, -1, 1) * pi_x^2 -sin(pi_y^2)),
  y = quote(runif(1, -1, 1) * pi_y^3-cos(pi_x^2))
 )
}
\seealso{
\code{\link{generate_plot}} the returned data frame is the input to generate the plot
}
