% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_filename.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate filenames}
\usage{
generate_filename(seed, filetype)
}
\arguments{
\item{seed}{a numeric input generated with create_seeds()}

\item{filetype}{set the file type for the image}
}
\value{
character vector, example: "2018-10-02-20-22_seed_22.png"
}
\description{
This function creates file names dynamically. The file name depends on date, time and specific seed.
}
\examples{
generate_filename(0104)
}
\seealso{
\code{\link{generate_seeds}} where the seed input is randomly generated
}
\keyword{cats}
