% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_img.R
\name{generate_img}
\alias{generate_img}
\title{Generate multiple generative images}
\usage{
generate_img(formula, nr_of_img, polar = FALSE, filetype = "png", ...)
}
\arguments{
\item{formula}{the formula you want to use as a list}

\item{nr_of_img}{the number of images that should be created}

\item{polar}{logical should the plot have a polar coordinate system ("polar = TRUE") or a cartesian coordinate system ("polar = FALSE")}

\item{filetype}{filetype of the final image. Default is `png`, for other options see the `devics` argument in `gggplot::gsave()`}

\item{...}{further arguments like color and background_color}
}
\value{
as many png files as you net in "nr_of_img"
}
\description{
This is the main function of the package generativeimgR. It calls all the other function neccessary to produce multiple generative images at once.
}
\examples{
generate_img(formula = my_formula, nr_of_img = 3, polar = FALSE)
generate_img(formula = my_formula, nr_of_img = 3, polar = FALSE, color = "#101820", background_color = "#F2AA4C")
}
\seealso{
\code{\link{generate_seeds}} generate the seeds for the randomness

\code{\link{generate_filename}} generate the file names

\code{\link{check_logfile_existence}} create a log file, if there is none

\code{\link{generate_logfile_entry}} generate the specific entry for the log file

\code{\link{generate_data}} generate the data depending on the formula

\code{\link{generate_plot}} plot the data and save a png file
}
