% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_logfile_entry.R
\name{generate_logfile_entry}
\alias{generate_logfile_entry}
\title{Generate the generative image}
\usage{
generate_logfile_entry(logfile, formula, seed, file_name)
}
\arguments{
\item{logfile}{the logfile}

\item{formula}{a string of the formula, that is used}

\item{seed}{the specific seed used}

\item{file_name}{the file name of the image}
}
\value{
a data frame and a saved csv-file
}
\description{
This function saves the information in order to recreate an image with random elements and appends it to the logfile.
}
\examples{
generate_logfile_entry(logfile, formula, seed, file_name)
}
\seealso{
\code{\link{check_logfile_existence}} check, if there is a logfile
}
