% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate one generative plot}
\usage{
generate_plot(df, file_name, polar, filetype)
}
\arguments{
\item{df}{the data frame created with `generate_data()`}

\item{file_name}{the file name for saving the plot}

\item{polar}{do you want to use a polar coordinate system? The default is a cartesian coordinate system.}

\item{file_name}{filetype of the final image. Default is `png`, for other options see the `devics` argument in `ggsave()`}
}
\value{
a png file
}
\description{
This function plots previously created data.
}
\examples{
generate_plot(df, file_name, polar = FALSE)
}
\seealso{
\code{\link{generate_data}} where the data is created
}
