% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formula_from_logfile.R
\name{get_formula_from_logfile}
\alias{get_formula_from_logfile}
\title{Get a formula from the logfile}
\usage{
get_formula_from_logfile(seed_to_recreate, filetype)
}
\arguments{
\item{seed_to_recreate}{the seed of the original image that either put in as a numeric value or vector from "get_seed_from_logfile()"}

\item{filetype}{set the file type for the image}
}
\value{
a list containing the formula
}
\description{
Extracts the formula of an image from the log file in order to be able to recreate the image.
}
\examples{
get_formula_from_logfile(0104)
}
\seealso{
\code{\link{get_seed_from_logfile}} get the seed, which is used as input

\code{\link{regenerate_img}} function to recreate an image
}
