% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed_from_logfile.R
\name{get_seed_from_logfile}
\alias{get_seed_from_logfile}
\title{Get one or several seeds from the logfile}
\usage{
get_seed_from_logfile(seed_to_recreate)
}
\arguments{
\item{seed_to_recreate}{the seed of the original image that either put in as a numeric value or vector from "get_seed_from_logfile()"}
}
\value{
a vector
}
\description{
Get one or several seeds from the logfile
}
\examples{
get_seed_from_logfile(0104)
}
\seealso{
\code{\link{get_formula_from_logfile}} get the formula for recreation

\code{\link{regenerate_img}} function to recreate an image
}
