% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasca.R
\name{my_ca}
\alias{my_ca}
\title{Run an alternate cellular automation algorithm on a matrix}
\usage{
my_ca(mat, iterations)
}
\arguments{
\item{mat}{a matrix}

\item{iterations}{number of iterations of the algorithm}
}
\value{
resulting matrix from the algorithm
}
\description{
Run an alternate cellular automation algorithm on a matrix
}
