% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenerate_img.R
\name{regenerate_img}
\alias{regenerate_img}
\title{Recreate an image}
\usage{
regenerate_img(seed_to_recreate, filetype, polar = FALSE)
}
\arguments{
\item{seed_to_recreate}{the seed of the original image that either put in as a numeric value or vector from "get_seed_from_logfile()"}

\item{filetype}{set the file type for the image}

\item{polar}{logical should the plot have a polar coordinate system ("polar = TRUE") or a cartesian coordinate system ("polar = FALSE")}
}
\value{
an image as png
}
\description{
This is the main function to recreate a previous generated image. It takes all relevant information from the log file to get exactly the same result. By using the same formula and the same seed you get the same randon numbers.
}
\examples{
regenerate_img(0104, polar = FALSE)
}
\seealso{
\code{\link{get_formula_from_logfile}} get the formula for recreation

\code{\link{get_seed_from_logfile}} get the seed, which is used as input
}
